Got all existing tests working
