+++ test/controllers/legal_internal/assignees_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class Admin::AssigneesControllerTest < ActionController::TestCase
+class LegalInternal::AssigneesControllerTest < ActionController::TestCase
   attr_reader :user, :application, :lawyer
 
   setup do
@@ -9,7 +9,7 @@ class Admin::AssigneesControllerTest < ActionController::TestCase
     @lawyer = users(:lawyer)
 
     sign_in_as(@user)
-    request.env["HTTP_REFERER"] = "http://test.host/admin/applications/#{@application.id}"
+    request.env["HTTP_REFERER"] = "http://test.host/legal_internal/applications/#{@application.id}"
   end
 
   test "POST #create associates lawyer to application and makes primary" do
@@ -21,15 +21,6 @@ class Admin::AssigneesControllerTest < ActionController::TestCase
 
     assert_redirected_to :back
     assert_equal 'Successfully assigned lawyer/student.', flash[:notice]
-  end
-
-  test "POST #create marks application as in_progress if it is pending_lawyer_match" do
-    application.state = "pending_lawyer_match"
-    application.save
-
-    post :create, assignee: { application_id: application.id, user_id: lawyer.id }
-
-    assert application.reload.in_progress?
   end
 
   test "POST #create handles invalid lawyer" do
